# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Student APP API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/degrees": {
            "get": {
                "tags": [
                    "Degrees"
                ],
                "summary": "List all degrees",
                "operationId": "050024f82119ab993e0af92190097ae0",
                "parameters": [
                    {
                        "name": "includeClassrooms",
                        "in": "query",
                        "description": "Whether to include the classrooms of each degree",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "includeStudents",
                        "in": "query",
                        "description": "Whether to include the students in each classroom",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of degrees",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DegreeResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Degrees"
                ],
                "summary": "Create a new degree",
                "operationId": "7362f1deba7f3f9fd4312d346b58dfc5",
                "requestBody": {
                    "description": "Degree data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "max_year"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "max_year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The created degree",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DegreeResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/v1/degrees/{degree}": {
            "get": {
                "tags": [
                    "Degrees"
                ],
                "summary": "Retrieve a single degree",
                "operationId": "da3ab9214624c025c3c084ffcf049832",
                "parameters": [
                    {
                        "name": "degree",
                        "in": "path",
                        "description": "ID of the degree to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "includeClassrooms",
                        "in": "query",
                        "description": "Whether to include the classroom for the degree",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "includeStudents",
                        "in": "query",
                        "description": "Whether to include the students in the classroom of the degree",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested degree",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DegreeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Degree not found"
                    }
                }
            }
        },
        "/api/v1/degrees/{id}": {
            "put": {
                "tags": [
                    "Degrees"
                ],
                "summary": "Update a degree with PUT request",
                "operationId": "4bc65057c0643c9aaa2f2751155c2fe2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the degree to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Degree data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "max_year"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "max_year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Degree updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DegreeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Degree not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Degrees"
                ],
                "summary": "Delete a degree",
                "operationId": "efd90cd652649bebbf237e85b8d0a14b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the degree to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "forceDelete",
                        "in": "query",
                        "description": "Whether to force delete the degree and its related records",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Degree deleted successfully"
                    },
                    "404": {
                        "description": "Degree not found"
                    },
                    "422": {
                        "description": "Cannot delete degree with attached classrooms"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Degrees"
                ],
                "summary": "Update a degree with PATCH request",
                "operationId": "e5a04cc80922644bf9fa56b9c6b3d1cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the degree to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Degree data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "max_year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Degree updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DegreeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Degree not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DegreeResource": {},
...
        }
    }
}